finished smooth acceleration started fixing heading corrections for automous

+++ src/main/include/commands/DriveByDistance.h
@@ -35,5 +35,5 @@ class DriveByDistance
   double initial_heading_;
   double final_distance_;
   double initial_distance_;
-  int max_speed_;
+  double max_speed_;
 };
